The randomQuery Function completed. Test for the randomQuery can be found in the testJavaKnowlege() function. It is expected that we will comment out the testJavaKnowelege function when we no long need testing for our tests.
The testRandomURLs() is ... I'll say complete.  But as the other random functions get written it will need to be vetted.
In the end the random test was written with try catch statements and an assert at the end when the tests fails.
There is a number of flags for test data that is used during the initial testing.